refactor(cli): extract format/types into dedicated modules; split cli.ts by subcommand (#179, #180)
Per #179 the CLI-layer file owned TranscribeResult and four output
formatters that lib.ts re-exported as part of the public \`./core\`
API — a layering violation. Per #180 cli.ts had grown to ~500 LOC
handling install, say, status, main, argument dispatch, star-prompt
gating, and output formatting all in one file.

## Files added

- \`src/types.ts\` — owns \`TranscribeResult\` (the public output
  shape). lib.ts re-exports from here instead of cli.ts. No more
  "public API points at a CLI file" layering violation.
- \`src/format.ts\` — \`formatTextOutput\` /
  \`formatJsonOutput\` / \`formatVerboseOutput\` /
  \`formatTranscriptOutput\` moved out of cli.ts.
- \`src/cli/install.ts\` — installCommand + performInstall +
  askForStar + resolveBackendFlag.
- \`src/cli/say.ts\` — sayCommand + resolveText + autoRouteVoice +
  pickVoiceForLang.
- \`src/cli/status.ts\` — statusCommand (12 lines, just wraps
  showStatus).
- \`src/cli/main.ts\` — mainCommand + MainCommandArgs +
  detectLanguage + checkLanguageMismatch.
- \`src/cli/dispatch.ts\` — runCli + SUBCOMMANDS + suggest-command
  fallback.

## Files modified

- \`src/cli.ts\` — now a 25-line re-export shim (was 491). Keeps the
  shebang + \`import.meta.main\` → \`runCli()\` entry point so
  \`bin/kesha.js\` keeps working, and re-exports everything the test
  suite and lib.ts imported from \`./cli\` so no import paths need
  updating anywhere else.
- \`src/lib.ts\` — \`TranscribeResult\` re-export now points at
  \`./types\` (was \`./cli\`). No more circular-import defense comment
  needed since the value lives in a layer neither \`./cli\` nor
  \`./lib\` imports from.

## Verified

- \`bunx tsc --noEmit\` clean
- \`bun test tests/unit\` — 109/109 green (zero test file changes)
- \`bun run src/cli.ts --help|--version|say --help|install --help\` all
  render correctly
- \`bun run src/cli.ts --json --toon a.ogg\` still exits 2 with mutex
  error (the runtime dispatch still routes to mainCommand)
- Public API smoke: \`import { toToon, type TranscribeResult,
  downloadModel, say } from \"@drakulavich/kesha-voice-kit/core\"\`
  resolves + toToon round-trips

## Scope decisions

- Kept \`detectLanguage\` + \`checkLanguageMismatch\` inside
  \`src/cli/main.ts\` rather than hoisting to a shared module. They're
  only used by mainCommand; no current caller in install/say/status.
- Kept the \`pkg.json\` read in \`src/cli/main.ts\` (for version
  metadata) and \`src/cli/install.ts\` (for star-prompt gating). Two
  JSON reads per CLI invocation; trivial.
- SUBCOMMANDS is still a hand-written array in dispatch.ts rather than
  citty's native \`subCommands\` option — the current structure runs
  each subcommand through \`runMain(...)\` so it exits cleanly; the
  native citty path is a separate concern.

Closes #179.
Closes #180.
